travis build scripts can run locally now.

+++ .travis/travis_build_linux.sh
@@ -42,11 +42,12 @@ if type $PYTHON3 > /dev/null; then $PYTHON3 -m compileall -q . ; fi
 
 echo "Currently in `pwd`"
 (
+    rm -rf /tmp/moose/usr && mkdir -p /tmp/moose/usr/{bin,lib}
     mkdir -p _GSL_BUILD && cd _GSL_BUILD
     cmake -DDEBUG=ON -DPYTHON_EXECUTABLE="$PYTHON2" \
         -DCMAKE_INSTALL_PREFIX=/tmp/moose/usr ..
     $MAKE && ctest --output-on-failure
-    sudo make install
+    make install
     PYTHONPATH=/tmp/moose/usr/lib/python2.7/site-packages $PYTHON2 -c 'import moose;print(moose.__file__);print(moose.version())'
 )
 
@@ -60,8 +61,10 @@ echo "Currently in `pwd`"
 # This is only applicable on linux build.
 echo "Python3: Removed python2-networkx and install python3"
 if type $PYTHON3 > /dev/null; then
-    sudo apt-get remove -qq python-networkx
-    sudo apt-get install -qq python3-networkx
+    if type apt-get --help > /dev/null; then
+        sudo apt-get remove -qq python-networkx
+        sudo apt-get install -qq python3-networkx
+    fi
     (
         mkdir -p _GSL_BUILD2 && cd _GSL_BUILD2 && \
             cmake -DDEBUG=ON -DPYTHON_EXECUTABLE="$PYTHON3" ..
